# internal/docs/swagger.yaml
definitions:
...
  github_com_umar5678_go-backend_internal_models.UserRole:
    enum:
    - rider
    - driver
    - admin
    - delivery_person
    - service_provider
    - handyman
    type: string
    x-enum-varnames:
    - RoleRider
    - RoleDriver
    - RoleAdmin
    - RoleDeliveryPerson
    - RoleServiceProvider
    - RoleHandyman
  github_com_umar5678_go-backend_internal_models.UserStatus:
    enum:
    - active
    - suspended
    - banned
    - pending_verification
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusSuspended
    - StatusBanned
    - StatusPendingVerification
...
  github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      name:
        type: string
      phone:
        type: string
      profilePhotoUrl:
        type: string
      role:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserStatus'
    type: object
...
  github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse:
    properties:
      acceptanceRate:
        type: number
      cancellationRate:
        type: number
      createdAt:
        type: string
      currentLocation:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.LocationResponse'
      heading:
        type: integer
      id:
        type: string
      isVerified:
        type: boolean
      licenseNumber:
        type: string
      rating:
        type: number
      status:
        type: string
      totalEarnings:
        type: number
      totalTrips:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      userId:
        type: string
      vehicle:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.VehicleResponse'
      walletBalance:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.LocationResponse:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  github_com_umar5678_go-backend_internal_modules_drivers_dto.VehicleResponse:
    properties:
      capacity:
        type: integer
      color:
        type: string
      createdAt:
        type: string
      driverId:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      licensePlate:
        type: string
      make:
        type: string
      model:
        type: string
      updatedAt:
        type: string
      vehicleType:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse'
      vehicleTypeId:
        type: string
      year:
        type: integer
    type: object
...
  github_com_umar5678_go-backend_internal_modules_todo_dto.CreateTodoRequest:
    properties:
      description:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - title
    type: object
  github_com_umar5678_go-backend_internal_modules_todo_dto.TodoResponse:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_todo_dto.UpdateTodoRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        minLength: 3
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_tracking_dto.DriverLocationResponse:
    properties:
      distance:
        description: km from search point
        type: number
      driver:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse'
      driverId:
        type: string
      eta:
        description: seconds
        type: integer
      location:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.LocationResponse'
    type: object
  github_com_umar5678_go-backend_internal_modules_tracking_dto.LocationResponse:
    properties:
      accuracy:
        type: number
      heading:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      speed:
        type: number
      timestamp:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_tracking_dto.NearbyDriversResponse:
    properties:
      count:
        type: integer
      drivers:
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.DriverLocationResponse'
        type: array
      radiusKm:
        type: number
      searchLocation:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.LocationResponse'
    type: object
  github_com_umar5678_go-backend_internal_modules_tracking_dto.UpdateLocationRequest:
    properties:
      accuracy:
        minimum: 0
        type: number
      heading:
        maximum: 360
        minimum: 0
        type: integer
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      speed:
        maximum: 300
        minimum: 0
        type: number
    required:
    - latitude
    - longitude
    type: object
  github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse:
    properties:
      baseFare:
        type: number
      bookingFee:
        type: number
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      displayName:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      perKmRate:
        type: number
      perMinuteRate:
        type: number
    type: object
...
  github_com_umar5678_go-backend_internal_utils_response.ErrorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Meta:
    properties:
      durationMs:
        type: number
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Response:
    properties:
      code:
        type: string
      data: {}
      errors:
        description: Use ErrorDetail
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.ErrorDetail'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Meta'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready Go backend API
  title: supr booking server in go
  version: "1.0"
paths:
...
  /todos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.TodoResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Todo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.TodoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create new todo
      tags:
      - todos
  /todos/{id}:
    delete:
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Delete todo
      tags:
      - todos
    get:
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.TodoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Todo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_todo_dto.TodoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update todo
      tags:
      - todos
  /tracking/driver/{driverId}:
    get:
      parameters:
      - description: Driver ID
        in: path
        name: driverId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.LocationResponse'
              type: object
      summary: Get driver's current location
      tags:
      - tracking
  /tracking/location:
    post:
      consumes:
      - application/json
      parameters:
      - description: Location data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Update driver location (called by driver app)
      tags:
      - tracking
  /tracking/nearby:
    get:
      parameters:
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Search radius in km
        in: query
        name: radiusKm
        type: number
      - description: Filter by vehicle type
        in: query
        name: vehicleTypeId
        type: string
      - description: Maximum results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_tracking_dto.NearbyDriversResponse'
              type: object
      summary: Find drivers near a location
      tags:
      - tracking
...
  /vehicles/types/{id}:
    get:
      parameters:
      - description: Vehicle Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse'
              type: object
      summary: Get vehicle type by ID
      tags:
      - vehicles
...
swagger: "2.0"